web: Add options to change input.
The input can now be changed via web. It should actually
work with siri too. The input settings are:
	input1 -> iphone dock
	input9 -> dvd
	tv     -> tv

I also had to adjust the button width because there are only
three input modes. It now becomes a part of the send_command
function. The function now looks too hacked...